requirements.txt
datasets==3.3.2
numpy==2.3.5
optuna==4.6.0
pandas==2.3.3
Pillow==12.0.0
plotly==6.5.0
PyYAML==6.0.3
torch==2.8.0+cu126
tqdm==4.67.1
transformers==4.57.0